# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VelesDB API
  description: High-performance vector database for AI applications. Supports semantic search, HNSW indexing, and multiple distance metrics. Authentication is optional — when API keys are configured via VELESDB_API_KEYS, all endpoints except /health and /ready require a valid Bearer token.
  contact:
    name: VelesDB Team
    url: https://github.com/cyberlife-coder/VelesDB
  license:
    name: VelesDB Core License 1.0
    url: https://github.com/cyberlife-coder/VelesDB/blob/main/LICENSE
  version: 6.0.0
servers:
- url: /
  description: Local server
paths:
...
  /collections/{name}/empty:
    get:
      tags:
      - collections
      summary: Check if a collection is empty.
      operationId: is_empty
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Empty status
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /collections/{name}/points/stream:
    post:
      tags:
      - points
      summary: Stream upsert points using NDJSON.
      description: |-
        Accepts a `application/x-ndjson` body. Each line is one JSON-encoded point,
        whose `id` is an unsigned 64-bit integer: a line that is not blank and does
        not parse as one, a string `id` included, is counted in the response's
        `malformed` field.
        Points are accumulated into micro-batches and flushed via `upsert_bulk`.

        The response body includes a `network_errors` field: a non-zero value means
        the HTTP body stream was truncated (e.g., client disconnect or proxy error),
        and the server may have received fewer points than the client sent (M-7).
      operationId: stream_upsert_points
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        description: NDJSON stream with one point per line
        content:
          application/x-ndjson:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Stream processed
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /collections/{name}/stream/enable:
    post:
      tags:
      - points
      summary: Enable streaming ingestion on a collection.
      description: |-
        Spawns the background drain task so subsequent calls to
        `/collections/{name}/stream/insert` accept points instead of returning a
        `409 Conflict` "streaming not configured" error. Omitted body fields fall
        back to the engine defaults. Calling this again replaces the existing
        ingester (the old drain task is aborted).

        Returns `200 OK` on success and `404 Not Found` when the collection does
        not exist.
      operationId: enable_streaming
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableStreamingRequest'
        required: true
      responses:
        '200':
          description: Streaming enabled
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
  /collections/{name}/stream/insert:
    post:
      tags:
      - points
      summary: Stream-insert a single point via the bounded ingestion channel.
      description: |-
        Returns 202 Accepted on success, 429 Too Many Requests when the buffer is
        full (with `Retry-After: 1` header per RFC 7231), 503 Service Unavailable
        when the drain task has exited, and 404 when the collection is not found.

        This handler is `async` to satisfy Axum's handler contract; it does not
        perform any async I/O internally (the channel send is non-blocking).
      operationId: stream_insert
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamInsertRequest'
        required: true
      responses:
        '202':
          description: Point accepted into streaming buffer
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Streaming not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
        '429':
          description: Streaming buffer full — retry after 1 second
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Streaming drain task has exited — collection must be reconfigured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    EnableStreamingRequest:
      type: object
      description: |-
        Request body for the enable-streaming endpoint.

        All fields are optional; omitted fields fall back to the engine defaults
        (`buffer_size = 10000`, `batch_size = 128`, `flush_interval_ms = 50`).
      properties:
        batch_size:
          type:
          - integer
          - 'null'
          description: Number of points that trigger an immediate micro-batch flush.
          minimum: 0
        buffer_size:
          type:
          - integer
          - 'null'
          description: Capacity of the bounded ingestion channel (backpressure threshold).
          minimum: 0
        flush_interval_ms:
          type:
          - integer
          - 'null'
          format: int64
          description: Maximum time (ms) before a partial batch is flushed.
          minimum: 0
    ErrorResponse:
      type: object
      description: Error response.
      required:
      - error
      properties:
        code:
          type:
          - string
          - 'null'
          description: Structured error code (e.g., `"VELES-004"` for dimension mismatch).
        error:
          type: string
          description: Error message.
...
    StreamInsertRequest:
      type: object
      description: Request body for the streaming insert endpoint (single point).
      required:
      - id
      - vector
      properties:
        id:
          oneOf:
          - type: integer
            format: int64
            minimum: 0
          - type: string
            pattern: ^[0-9]+$
          description: Point ID. Accepts a JSON integer (native form) or a string; use a string for u64 values above 2^53-1 to avoid JavaScript precision loss.
        payload:
          description: Optional JSON payload (metadata).
        vector:
          type: array
          items:
            type: number
            format: float
          description: Dense vector data.
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...